+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changes
 - Enhanced feature attachment viewer now includes panoramic images support.
 - Ability to copy printing parameters as a URL for sharing.
 - Support for multiple replicas and long-runnning requests.
+- Ctrl+Click opens feature attachment in a fixed tab.
 
 
 4.6.0